scan on home page
