+++ src/hooks/useReceiptScanDrop.tsx
@@ -1,0 +1,121 @@
+import {useMemo} from 'react';
+import {navigateToParticipantPage} from '@libs/IOUUtils';
+import Navigation from '@libs/Navigation/Navigation';
+import {hasOnlyPersonalPolicies as hasOnlyPersonalPoliciesUtil, isPaidGroupPolicy} from '@libs/PolicyUtils';
+import {generateReportID, getPolicyExpenseChat, isSelfDM} from '@libs/ReportUtils';
+import {shouldRestrictUserBillableActions} from '@libs/SubscriptionUtils';
+import {setTransactionReport} from '@libs/actions/Transaction';
+import type {ReceiptFile} from '@pages/iou/request/step/IOURequestStepScan/types';
+import CONST from '@src/CONST';
+import ONYXKEYS from '@src/ONYXKEYS';
+import ROUTES from '@src/ROUTES';
+import type {Transaction} from '@src/types/onyx';
+import type {FileObject} from '@src/types/utils/Attachment';
+import {initMoneyRequest, setMoneyRequestParticipantsFromReport, setMoneyRequestReceipt} from '@userActions/IOU';
+import {buildOptimisticTransactionAndCreateDraft} from '@userActions/TransactionEdit';
+import useCurrentUserPersonalDetails from './useCurrentUserPersonalDetails';
+import useFilesValidation from './useFilesValidation';
+import useOnyx from './useOnyx';
+import useSelfDMReport from './useSelfDMReport';
+
+/**
+ * Encapsulates the receipt scan drag-and-drop logic used by SearchPage and HomePage.
+ * Returns the drop handler, PDF validation component, and error modal needed for drag-and-drop receipt scanning.
+ */
+function useReceiptScanDrop() {
+    const currentUserPersonalDetails = useCurrentUserPersonalDetails();
+    const selfDMReport = useSelfDMReport();
+    const [policies] = useOnyx(ONYXKEYS.COLLECTION.POLICY, {canBeMissing: true});
+    const [currentDate] = useOnyx(ONYXKEYS.CURRENT_DATE, {canBeMissing: true});
+    const [activePolicyID] = useOnyx(ONYXKEYS.NVP_ACTIVE_POLICY_ID, {canBeMissing: false});
+    const [activePolicy] = useOnyx(`${ONYXKEYS.COLLECTION.POLICY}${activePolicyID}`, {canBeMissing: true});
+    const [personalPolicyID] = useOnyx(ONYXKEYS.PERSONAL_POLICY_ID, {canBeMissing: true});
+    const [personalPolicy] = useOnyx(`${ONYXKEYS.COLLECTION.POLICY}${personalPolicyID}`, {canBeMissing: true});
+    const [draftTransactions] = useOnyx(ONYXKEYS.COLLECTION.TRANSACTION_DRAFT, {canBeMissing: true});
+
+    const newReportID = useMemo(() => generateReportID(), []);
+    const [newReport] = useOnyx(`${ONYXKEYS.COLLECTION.REPORT}${newReportID}`, {canBeMissing: true});
+    const [newParentReport] = useOnyx(`${ONYXKEYS.COLLECTION.REPORT}${newReport?.parentReportID}`, {canBeMissing: true});
+
+    const hasOnlyPersonalPolicies = useMemo(() => hasOnlyPersonalPoliciesUtil(policies), [policies]);
+
+    const saveFileAndInitMoneyRequest = (files: FileObject[]) => {
+        const initialTransaction = initMoneyRequest({
+            isFromGlobalCreate: true,
+            isFromFloatingActionButton: true,
+            reportID: newReportID,
+            personalPolicy,
+            newIouRequestType: CONST.IOU.REQUEST_TYPE.SCAN,
+            report: newReport,
+            parentReport: newParentReport,
+            currentDate,
+            currentUserPersonalDetails,
+            hasOnlyPersonalPolicies,
+            draftTransactions,
+        });
+
+        const newReceiptFiles: ReceiptFile[] = [];
+
+        for (const [index, file] of files.entries()) {
+            const source = URL.createObjectURL(file as Blob);
+            const transaction =
+                index === 0
+                    ? (initialTransaction as Partial<Transaction>)
+                    : buildOptimisticTransactionAndCreateDraft({
+                          initialTransaction: initialTransaction as Partial<Transaction>,
+                          currentUserPersonalDetails,
+                          reportID: newReportID,
+                      });
+            const transactionID = transaction.transactionID ?? CONST.IOU.OPTIMISTIC_TRANSACTION_ID;
+            newReceiptFiles.push({
+                file,
+                source,
+                transactionID,
+            });
+            setMoneyRequestReceipt(transactionID, source, file.name ?? '', true, file.type);
+        }
+
+        if (isPaidGroupPolicy(activePolicy) && activePolicy?.isPolicyExpenseChatEnabled && !shouldRestrictUserBillableActions(activePolicy.id)) {
+            const shouldAutoReport = !!activePolicy?.autoReporting || !!personalPolicy?.autoReporting;
+            const report = shouldAutoReport ? getPolicyExpenseChat(currentUserPersonalDetails.accountID, activePolicy?.id) : selfDMReport;
+            const transactionReportID = isSelfDM(report) ? CONST.REPORT.UNREPORTED_REPORT_ID : report?.reportID;
+            const iouTypeTrackOrSubmit = transactionReportID === CONST.REPORT.UNREPORTED_REPORT_ID ? CONST.IOU.TYPE.TRACK : CONST.IOU.TYPE.SUBMIT;
+            const setParticipantsPromises = newReceiptFiles.map((receiptFile) => {
+                setTransactionReport(receiptFile.transactionID, {reportID: transactionReportID}, true);
+                return setMoneyRequestParticipantsFromReport(receiptFile.transactionID, report, currentUserPersonalDetails.accountID);
+            });
+            Promise.all(setParticipantsPromises).then(() =>
+                Navigation.navigate(
+                    ROUTES.MONEY_REQUEST_STEP_CONFIRMATION.getRoute(
+                        CONST.IOU.ACTION.CREATE,
+                        iouTypeTrackOrSubmit,
+                        initialTransaction?.transactionID ?? CONST.IOU.OPTIMISTIC_TRANSACTION_ID,
+                        report?.reportID,
+                    ),
+                ),
+            );
+        } else {
+            navigateToParticipantPage(CONST.IOU.TYPE.CREATE, CONST.IOU.OPTIMISTIC_TRANSACTION_ID, newReportID);
+        }
+    };
+
+    const {validateFiles, PDFValidationComponent, ErrorModal} = useFilesValidation(saveFileAndInitMoneyRequest);
+
+    const initScanRequest = (e: DragEvent) => {
+        const files = Array.from(e?.dataTransfer?.files ?? []);
+
+        if (files.length === 0) {
+            return;
+        }
+        for (const file of files) {
+            // eslint-disable-next-line no-param-reassign
+            file.uri = URL.createObjectURL(file);
+        }
+
+        validateFiles(files, Array.from(e.dataTransfer?.items ?? []));
+    };
+
+    return {initScanRequest, PDFValidationComponent, ErrorModal};
+}
+
+export default useReceiptScanDrop;
